evaluators - related_resources, searchable ; moved licence helper methods from FAIRCheck to licence evaluator
