+++ fuji_server/evaluators/fair_evaluator_license.py
@@ -32,6 +32,52 @@ import idutils
 from fuji_server.helper.metadata_mapper import Mapper
 class FAIREvaluatorLicense(FAIREvaluator):
 
+    def isLicense (self, value, metric_id):
+        islicense = False
+        isurl = idutils.is_url(value)
+        spdx_html = None
+        spdx_osi = None
+        if isurl:
+            spdx_html, spdx_osi = self.lookup_license_by_url(value, metric_id)
+        else:
+            spdx_html, spdx_osi = self.lookup_license_by_name(value, metric_id)
+        if spdx_html or spdx_osi:
+            islicense = True
+        return islicense
+
+    def lookup_license_by_url(self, u, metric_id):
+        self.logger.info('{0} : Verify URL through SPDX registry - {1}'.format(metric_id, u))
+        html_url = None
+        isOsiApproved = False
+        for item in self.fuji.SPDX_LICENSES:
+            # u = u.lower()
+            # if any(u in v.lower() for v in item.values()):
+            seeAlso = item['seeAlso']
+            if any(u in v for v in seeAlso):
+                self.logger.info('{0} : Found SPDX license representation - {1}'.format(metric_id, item['detailsUrl']))
+                # html_url = '.html'.join(item['detailsUrl'].rsplit('.json', 1))
+                html_url = item['detailsUrl'].replace(".json", ".html")
+                isOsiApproved = item['isOsiApproved']
+                break
+        return html_url, isOsiApproved
+
+    def lookup_license_by_name(self, lvalue, metric_id):
+        # TODO - find simpler way to run fuzzy-based search over dict/json (e.g., regex)
+        html_url = None
+        isOsiApproved = False
+        self.logger.info('{0} : Verify name through SPDX registry - {1}'.format(metric_id, lvalue))
+        # Levenshtein distance similarity ratio between two license name
+        sim = [Levenshtein.ratio(lvalue.lower(), i) for i in self.fuji.SPDX_LICENSE_NAMES]
+        if max(sim) > 0.85:
+            index_max = max(range(len(sim)), key=sim.__getitem__)
+            sim_license = self.fuji.SPDX_LICENSE_NAMES[index_max]
+            found = next((item for item in self.fuji.SPDX_LICENSES if item['name'] == sim_license), None)
+            self.logger.info('{0}: Found SPDX license representation - {1}'.format(metric_id,found['detailsUrl']))
+            # html_url = '.html'.join(found['detailsUrl'].rsplit('.json', 1))
+            html_url = found['detailsUrl'].replace(".json", ".html")
+            isOsiApproved = found['isOsiApproved']
+        return html_url, isOsiApproved
+
     def evaluate(self):
 
         self.result = License(id=self.count, metric_identifier=self.metric_identifier, metric_name=self.metric_name)
@@ -48,9 +94,9 @@ class FAIREvaluatorLicense(FAIREvaluator):
                 if isinstance(l, str):
                     isurl = idutils.is_url(l)
                 if isurl:
-                    spdx_html, spdx_osi = self.fuji.lookup_license_by_url(l, self.metric_identifier)
+                    spdx_html, spdx_osi = self.lookup_license_by_url(l, self.metric_identifier)
                 else:  # maybe licence name
-                    spdx_html, spdx_osi = self.fuji.lookup_license_by_name(l, self.metric_identifier)
+                    spdx_html, spdx_osi = self.lookup_license_by_name(l, self.metric_identifier)
                 if not spdx_html:
                     self.logger.warning('FsF-R1.1-01M : NO SPDX license representation (spdx url, osi_approved) found')
                 license_output.details_url = spdx_html
